[FEAT] 패키지 구조 정리: 클린 아키텍처 + 레이어 아키텍처 적용
