--- src/main/java/kr/hhplus/be/server/global/response/Code.java
@@ -1,6 +1,0 @@
-package kr.hhplus.be.server.global.response;
-
-public interface Code {
-    String getCode();
-    String getMessage();
-}
