+++ src/main/java/kr/hhplus/be/server/interfaces/common/Code.java
@@ -1,0 +1,6 @@
+package kr.hhplus.be.server.interfaces.common;
+
+public interface Code {
+    String getCode();
+    String getMessage();
+}
