+++ src/main/java/kr/hhplus/be/server/interfaces/product/ProductController.java
@@ -1,0 +1,43 @@
+package kr.hhplus.be.server.interfaces.product;
+
+import kr.hhplus.be.server.interfaces.common.ApiResult;
+import kr.hhplus.be.server.interfaces.common.SuccessCode;
+import kr.hhplus.be.server.util.FakeStore;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+@RequiredArgsConstructor
+@RestController
+@RequestMapping("/products")
+public class ProductController implements ProductApi {
+
+    private final FakeStore fakeStore;
+
+    @Override
+    @GetMapping("/{id}")
+    public ResponseEntity<ApiResult<ProductResponse.Detail>> findProduct (
+            @PathVariable("id") Long productId
+    ) {
+        ProductResponse.Detail response = fakeStore.product();
+        return ResponseEntity
+                .status(HttpStatus.OK)
+                .body(ApiResult.of(SuccessCode.FIND_PRODUCT, response));
+    }
+
+    @Override
+    @GetMapping("/popular")
+    public ResponseEntity<ApiResult<List<ProductResponse.Popular>>> findPopularProducts() {
+        List<ProductResponse.Popular> response = fakeStore.popularProducts();
+        return ResponseEntity
+                .status(HttpStatus.OK)
+                .body(ApiResult.of(SuccessCode.FIND_POPULAR_PRODUCT, response));
+    }
+
+}
